Don't retry IDE server connection on some response
After starting the server we attempt to load modules and retry in case the server doesn't
seem to be started. But the server may be replying "no output directory" or other issue
loading modules, so only retry if there is some error connecting or parsing the result
and not a valid error response.

Fix #48